% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BMData}
\alias{BMData}
\title{A list of eight Benchmark data sets}
\format{A list with eight elements}
\usage{
data(BMData)
}
\description{
A list of eight Benchmark data sets. Input variable is u, and output variable is y
}
\details{
\describe{
   \item{BC}{BC,Base Configuration}
   \item{SD}{SD,Squaring device}
   \item{FL}{FL,Feedback loop}
   \item{CN}{CN,Colored noise}
   \item{SF}{SF,Step Function}
   \item{FT}{FT,Feed Through}
   \item{DT}{Dt,Dead Time}
   \item{SS}{SS,second-order system}
}
}
\keyword{datasets}

