% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.R
\name{CCF}
\alias{CCF}
\title{Calculate the Cross Correlation Function}
\usage{
CCF(x, y, lag.max, ...)
}
\arguments{
\item{x}{A vector represents a time sequence}

\item{y}{A vector represents another time sequence}

\item{lag.max}{A numeric determines the max time lag}
}
\value{
A dataframe of time lag and its cross correlation value
}
\description{
Takes in three objects: x, y and lag.max. x and y is a vector, lag.max is a numeric
}

