% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCM.R
\name{CCM}
\alias{CCM}
\title{Calculate the CCM of two sequences}
\usage{
CCM(x, y, maxLag = 20, timestep = 1, tag, E = 2, ...)
}
\arguments{
\item{x}{A vector represents the time sequence}

\item{y}{A vector represents another time sequence}

\item{maxLag}{Max time lag needs to be calculated}

\item{timestep}{Interval of time lag}

\item{tag}{Select the dots of x and y to calculate the CCM}

\item{E+1}{The embedded dimension of the manifold}
}
\value{
Return a list of time lag sequence, yCausex and xCausey
}
\description{
Takes in two sequences, calculate CCM value in different time lag, and plot the result
}
\examples{
set.seed(1)
y<-rnorm(100)
x<-y*2+1
tag=c(1:100)
result<-CCM(x,y,maxLag=20,timestep=1,tag=tag,E=2)
# plot the data
testData<-data.frame(u,y)
matplot (testData, type = "l", lty = 1:2, lwd = 1, pch = NULL,
         col = 1:2)
legend("topright",colnames(testData),lty = 1:2, lwd = 1, pch = NULL,
      col = 1:2)
}

