% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cstr}
\alias{cstr}
\title{Process data of a simulating CSTR}
\format{A data frame with 99991 rows and 10 variables}
\usage{
data(cstr)
}
\description{
A dataset containing the material concentration, temperature and time of 99991.
\describe{
   \item{Ca}{Concentration of A, in mol/L}
   \item{Cb}{Concentration of B, in mol/L}
   \item{Cc}{Concentration of C, in mol/L}
   \item{Ci}{Input Concentration of A, 1 mol/L with white noise having N(0,0.1)}
   \item{Tf}{Reaction Temperature, 350K with white noise having N(0,3)}
   \item{time}{time, start from 1 second, sample interval is 0.1 second}
}
}
\details{
The model of CSTR is as followed.

dCa(t)=F/V*(Ci(t)-Ca(t))-k1*Ca(t)*exp(-E1/(R*Tf))

dCb(t)=k1*Ca(t)*exp(-E1/(R*Tf))-k2*Cb(t)*exp(-E2/(R*Tf))-F/V*Cb(t)

Cc(t)=k2*Cb(t)*exp(-E2/(R*Tf))-F/V*Cc(t)
}
\keyword{data.frame}

