% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate the Euclidean distance between two matrice}
\usage{
euclidean(A, B, ...)
}
\arguments{
\item{A}{A matrix or a vector}

\item{B}{Another matrix or another vector}
}
\value{
The distance between two matrice
}
\description{
Takes in two matrice and return the Euclidean distance
}

