% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GT.R
\name{GT}
\alias{GT}
\title{To do the granger test between two time series}
\usage{
GT(x, y, order)
}
\arguments{
\item{x}{A vector represents one variable}

\item{y}{A vector represents another variable}

\item{order}{a number indicating the time order in line model}
}
\value{
A dataframe including sic items: order, EU_Y is the sum of residuals squares in restricted model,
aicEU_Y is the AIC value in restricted model, EUY is the sum of residuals squares in unrestricted model,
aicEUY is the AIC value in unrestricted model, dw is the Durbin-Watson test value.
}
\description{
To do the granger test between two time series
}
\details{
The model used in this function is just linear model(lm), the way to estimate the linear model is
OLS. To consider the model is consistent, dw value should be between 2 and 3.
}

