% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{Find k nearest neighbors}
\usage{
knn(x, i, dist = euclidean, k = 2, ...)
}
\arguments{
\item{x}{A matrix of all dots, each row represents a dot}

\item{i}{A numeric represents which row to calculate}

\item{dist}{The method to calculate the distance, default is euclidean distance}

\item{k}{A numeric represents the number of neighbors}
}
\value{
Return a list of k nearest distance and row number
}
\description{
Takes in a matrix and a number, k nearest neighbors
}

