% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiParaCCM.R
\name{multiParaCCM}
\alias{multiParaCCM}
\title{Calculate the CCM of two sequences in Parallel}
\usage{
multiParaCCM(x, y, maxLag = 20, timestep = 1, tag, E = 2, ...)
}
\arguments{
\item{x}{A vector represents the time sequence}

\item{y}{A vector represents another time sequence}

\item{maxLag}{Max time lag needs to be calculated}

\item{timestep}{Interval of time lag}

\item{tag}{Select the dots of x and y to calculate the CCM}

\item{E+1}{The embedded dimension of the manifold}
}
\value{
Return a list of time lag sequence, yCausex and xCausey
}
\description{
Takes in two sequences, calculate CCM value in different time lag, and plot the result.
The CCM value is mean value of corY.
}

