% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCCM.R
\name{myCCM}
\alias{myCCM}
\title{Calculate the CCM value of x and y}
\usage{
myCCM(x, y, lag, tag, E = 2, k = 2, way = 2, ...)
}
\arguments{
\item{x}{A vector represents the time sequence}

\item{y}{A vector represents another time sequence}

\item{lag}{Time lag between x and y}

\item{tag}{Select the dots of x and y to calculate the CCM}

\item{k}{K nearest neighbors to construct the manifold}

\item{E+1}{The embedded dimension of the manifold}
}
\value{
A list of time lag and correlation number
}
\description{
Takes in two sequences, and calculate the value of CCM
}

