% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probEstimate.R
\name{probEstimate}
\alias{probEstimate}
\title{To calculate the kernel estimator of a vector}
\usage{
probEstimate(x)
}
\arguments{
\item{x}{A vector to estimate the density probability}
}
\value{
A dataframe of x and its probability
}
\description{
To calculate the kernel estimator of a vector
}
\details{
This function uses the Gaussian kernel function to estimate the probability
}

