% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standize.R
\name{standize}
\alias{standize}
\title{Standardization of a vector}
\usage{
standize(x, way = 2, ...)
}
\arguments{
\item{x}{A vector to be standized}

\item{way}{A number to determine which method to standize the vector.
way=1 means the normalization of vector.
way=2 means use the max and min to standize.}
}
\value{
The standardization of a vector
}
\description{
Takes in a vector and a number
}

