% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGT.R
\name{testGT}
\alias{testGT}
\title{To test the result of granger test}
\usage{
testGT(x, y, order.max)
}
\arguments{
\item{x}{A vector represents one variable}

\item{y}{A vector represents another variable}

\item{order.amx}{The max time lag in granger test.}
}
\value{
A dataframe including sic items: judge is the judgement of causality,order, EU_Y is the sum of
residuals squares in restricted model,
aicEU_Y is the AIC value in restricted model, EUY is the sum of residuals squares in unrestricted model,
aicEUY is the AIC value in unrestricted model, dw is the Durbin-Watson test value.
}
\description{
To test the result of granger test
}
\details{
There is two test for result. First test is to test whether the model is consistent, i.e. the dw is between
 2 and 3. Second test is to judge whether the sum of residuals squares is different.
}
\examples{
data("BMData")
i=7
x<-BMData[[i]]$u
y<-BMData[[i]]$y
result<-testGT(x,y,order.max=20)
}

